Move .env validation to separate file. Update migrations: get migrationNames by map() function
